+++ blueprints/front/static/uploads/img_plantilla/priorizacion.jpg
