+++ blueprints/front/templates/home.html
